rewrite fault logging system with timestamp, severity, category; use hardware to check OV, UV flags

+++ ADBMS6830/program/inc/serialPrintResult.h
@@ -37,7 +37,7 @@ void printDiagnosticTestResult(uint8_t tIC, cell_asic *IC, DIAGNOSTIC_TYPE type)
 void diagnosticTestResultPrint(uint8_t result);
 void printOpenWireTestResult(uint8_t tIC, cell_asic *IC, TYPE type);
 void openWireResultPrint(uint8_t result);
-float getVoltage(int data);
+float getVoltage(uint16_t data);
 void printPollAdcConvTime(int count);
 void printMenu();
